Reader hits db in separate thread when resumed

+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderPostListFragment.java
@@ -226,33 +226,37 @@ public class ReaderPostListFragment extends Fragment
             AppLog.d(T.READER, "reader post list > resumed from paused state");
             mWasPaused = false;
 
-            // default to refreshing posts in case the user returned from an activity that
-            // changed one (or more) of them
-            boolean shouldRefreshPosts = true;
-
             if (getPostListType().equals(ReaderPostListType.TAG_FOLLOWED)) {
-                // check if the user added a tag in ReaderSubsActivity
-                Object event = EventBus.getDefault().getStickyEvent(ReaderEvents.TagAdded.class);
-                if (event != null) {
-                    String tagName = ((ReaderEvents.TagAdded) event).getTagName();
-                    EventBus.getDefault().removeStickyEvent(event);
-                    ReaderTag newTag = new ReaderTag(tagName, ReaderTagType.FOLLOWED);
-                    setCurrentTag(newTag, true);
-                    shouldRefreshPosts = false;
-                // make sure the current tag is still valid
-                } else if (!ReaderTagTable.tagExists(getCurrentTag())) {
-                    AppLog.d(T.READER, "reader post list > current tag no longer valid");
-                    setCurrentTag(ReaderTag.getDefaultTag(), true);
-                    shouldRefreshPosts = false;
-                // auto-update the current tag if it's time
-                } else if (!isUpdating() && ReaderTagTable.shouldAutoUpdateTag(getCurrentTag())) {
-                    AppLog.i(T.READER, "reader post list > auto-updating current tag after resume");
-                    updatePostsWithTag(getCurrentTag(), UpdateAction.REQUEST_NEWER);
-                    shouldRefreshPosts = false;
-                }
-            }
+                // use a separate thread since we're hitting the database here
+                new Thread() {
+                    @Override
+                    public void run() {
+                        // user may have deleted the current tag in ReaderSubsActivity
+                        final boolean isTagValid = ReaderTagTable.tagExists(getCurrentTag());
 
-            if (shouldRefreshPosts) {
+                        // if tag is still valid, check if it's time to auto-update its posts
+                        final boolean timeToAutoUpdate = isTagValid && ReaderTagTable.shouldAutoUpdateTag(getCurrentTag());
+
+                        getActivity().runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                if (!isTagValid) {
+                                    // current tag no longer exists, revert to default
+                                    AppLog.d(T.READER, "reader post list > current tag no longer valid");
+                                    setCurrentTag(ReaderTag.getDefaultTag(), true);
+                                } else if (timeToAutoUpdate && !isUpdating()) {
+                                    // auto-update the current tag if it's time
+                                    AppLog.i(T.READER, "reader post list > auto-updating current tag");
+                                    updateCurrentTag();
+                                } else {
+                                    // otherwise, refresh posts to make sure any changes are reflected
+                                    refreshPosts();
+                                }
+                            }
+                        });
+                    }
+                }.start();
+            } else {
                 refreshPosts();
             }
         }
@@ -332,7 +336,7 @@ public class ReaderPostListFragment extends Fragment
 
         Context context = container.getContext();
 
-        // add the item decoration (divivers) to the recycler, skipping the first item if the first
+        // add the item decoration (dividers) to the recycler, skipping the first item if the first
         // item is the tag toolbar (shown when viewing posts in followed tags) - this is to avoid
         // having the tag toolbar take up more vertical space than necessary
         int spacingHorizontal = context.getResources().getDimensionPixelSize(R.dimen.reader_card_margin);
@@ -1042,7 +1046,9 @@ public class ReaderPostListFragment extends Fragment
      * since we don't want to purge posts that the user would expect to see when offline
      */
     private void purgeDatabaseIfNeeded() {
-        if (!mHasPurgedReaderDb && NetworkUtils.isNetworkAvailable(getActivity())) {
+        if (!mHasPurgedReaderDb
+                && isAdded()
+                && NetworkUtils.isNetworkAvailable(getActivity())) {
             AppLog.d(T.READER, "reader post list > purging database");
             mHasPurgedReaderDb = true;
             ReaderDatabase.purgeAsync();
